requirements.txt
django           
django-mdeditor  
Markdown        
gunicorn
django-heroku